Seperate AssetManifest subclasses into own files
